Created separate lighthouse config file
