+++ .github/workflows/dashboards-lighthouse-test.yml
@@ -59,7 +59,7 @@ jobs:
           start: npx cross-env NODE_PATH=${GITHUB_WORKSPACE}/node_modules node node_modules/@highcharts/highcharts-utils/server --localOnly
           wait-on: 'http://localhost:3031/dashboards.js'
           env: type=base
-          config: specPattern=**/performance/*.cy.js
+          config-file: test/cypress/configs/dashboards.lighthouse.config.mjs
           install: false
 
       - name: Stop server
@@ -95,7 +95,7 @@ jobs:
           start: npx cross-env NODE_PATH=${GITHUB_WORKSPACE}/node_modules node node_modules/@highcharts/highcharts-utils/server --localOnly
           wait-on: 'http://localhost:3031/dashboards.js'
           env: type=actual
-          config: specPattern=**/performance/*.cy.js
+          config-file: test/cypress/configs/dashboards.lighthouse.config.mjs
           install: false
 
       - name: Add to summary
